requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
requests==2.31.0
numpy==1.26.4
scikit-learn==1.4.2
python-dotenv==1.0.1
